Move ExerciseUpdater to another package
